deduplicate some git logic + simplify some functions
