+++ enterprise/internal/codeintel/gitserver/client.go
@@ -6,13 +6,13 @@ import (
 	"fmt"
 	"os"
 	"regexp"
-	"sort"
 	"strings"
 	"time"
 
 	"github.com/opentracing/opentracing-go/log"
 	"github.com/pkg/errors"
 
+	"github.com/sourcegraph/sourcegraph/enterprise/lib/codeintel/pathexistence"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
@@ -200,73 +200,13 @@ func (c *Client) DirectoryChildren(ctx context.Context, repositoryID int, commit
 	}})
 	defer endObservation(1, observation.Args{})
 
-	out, err := c.execResolveRevGitCommand(ctx, repositoryID, commit, append([]string{"ls-tree", "--name-only", commit, "--"}, cleanDirectoriesForLsTree(dirnames)...)...)
-	if err != nil {
-		return nil, err
-	}
-
-	return parseDirectoryChildren(dirnames, strings.Split(out, "\n")), nil
-}
-
-// cleanDirectoriesForLsTree sanitizes the input dirnames to a git ls-tree command. There are a
-// few peculiarities handled here:
-//
-//   1. The root of the tree must be indicated with `.`, and
-//   2. In order for git ls-tree to return a directory's contents, the name must end in a slash.
-func cleanDirectoriesForLsTree(dirnames []string) []string {
-	var args []string
-	for _, dir := range dirnames {
-		if dir == "" {
-			args = append(args, ".")
-		} else {
-			if !strings.HasSuffix(dir, "/") {
-				dir += "/"
-			}
-			args = append(args, dir)
-		}
-	}
-
-	return args
-}
-
-// parseDirectoryChildren converts the flat list of files from git ls-tree into a map. The keys of the
-// resulting map are the input (unsanitized) dirnames, and the value of that key are the files nested
-// under that directory. If dirnames contains a directory that encloses another, then the paths will
-// be placed into the key sharing the longest path prefix.
-func parseDirectoryChildren(dirnames, paths []string) map[string][]string {
-	childrenMap := map[string][]string{}
-
-	// Ensure each directory has an entry, even if it has no children
-	// listed in the gitserver output.
-	for _, dirname := range dirnames {
-		childrenMap[dirname] = nil
-	}
-
-	// Order directory names by length (biggest first) so that we assign
-	// paths to the most specific enclosing directory in the following loop.
-	sort.Slice(dirnames, func(i, j int) bool {
-		return len(dirnames[i]) > len(dirnames[j])
-	})
-
-	for _, path := range paths {
-		if strings.Contains(path, "/") {
-			for _, dirname := range dirnames {
-				if strings.HasPrefix(path, dirname) {
-					childrenMap[dirname] = append(childrenMap[dirname], path)
-					break
-				}
-			}
-		} else {
-			// No need to loop here. If we have a root input directory it
-			// will necessarily be the last element due to the previous
-			// sorting step.
-			if len(dirnames) > 0 && dirnames[len(dirnames)-1] == "" {
-				childrenMap[""] = append(childrenMap[""], path)
-			}
-		}
-	}
-
-	return childrenMap
+	return pathexistence.GitGetChildren(
+		func(args ...string) (string, error) {
+			return c.execResolveRevGitCommand(ctx, repositoryID, commit, args...)
+		},
+		commit,
+		dirnames,
+	)
 }
 
 // FileExists determines whether a file exists in a particular commit of a repository.
