+++ enterprise/internal/codeintel/gitserver/client_test.go
@@ -115,75 +115,3 @@ func TestParseCommitGraphPartial(t *testing.T) {
 		t.Errorf("unexpected commit order (-want +got):\n%s", diff)
 	}
 }
-
-func TestParseDirectoryChildrenRoot(t *testing.T) {
-	dirnames := []string{""}
-	paths := []string{
-		".github",
-		".gitignore",
-		"LICENSE",
-		"README.md",
-		"cmd",
-		"go.mod",
-		"go.sum",
-		"internal",
-		"protocol",
-	}
-
-	expected := map[string][]string{
-		"": paths,
-	}
-
-	if diff := cmp.Diff(expected, parseDirectoryChildren(dirnames, paths)); diff != "" {
-		t.Errorf("unexpected directory children result (-want +got):\n%s", diff)
-	}
-}
-
-func TestParseDirectoryChildrenNonRoot(t *testing.T) {
-	dirnames := []string{"cmd/", "protocol/", "cmd/protocol/"}
-	paths := []string{
-		"cmd/lsif-go",
-		"protocol/protocol.go",
-		"protocol/writer.go",
-	}
-
-	expected := map[string][]string{
-		"cmd/":          {"cmd/lsif-go"},
-		"protocol/":     {"protocol/protocol.go", "protocol/writer.go"},
-		"cmd/protocol/": nil,
-	}
-
-	if diff := cmp.Diff(expected, parseDirectoryChildren(dirnames, paths)); diff != "" {
-		t.Errorf("unexpected directory children result (-want +got):\n%s", diff)
-	}
-}
-
-func TestParseDirectoryChildrenDifferentDepths(t *testing.T) {
-	dirnames := []string{"cmd/", "protocol/", "cmd/protocol/"}
-	paths := []string{
-		"cmd/lsif-go",
-		"protocol/protocol.go",
-		"protocol/writer.go",
-		"cmd/protocol/main.go",
-	}
-
-	expected := map[string][]string{
-		"cmd/":          {"cmd/lsif-go"},
-		"protocol/":     {"protocol/protocol.go", "protocol/writer.go"},
-		"cmd/protocol/": {"cmd/protocol/main.go"},
-	}
-
-	if diff := cmp.Diff(expected, parseDirectoryChildren(dirnames, paths)); diff != "" {
-		t.Errorf("unexpected directory children result (-want +got):\n%s", diff)
-	}
-}
-
-func TestCleanDirectoriesForLsTree(t *testing.T) {
-	args := []string{"", "foo", "bar/", "baz"}
-	actual := cleanDirectoriesForLsTree(args)
-	expected := []string{".", "foo/", "bar/", "baz/"}
-
-	if diff := cmp.Diff(expected, actual); diff != "" {
-		t.Errorf("unexpected ls-tree args (-want +got):\n%s", diff)
-	}
-}
