+++ enterprise/lib/codeintel/semantic/repl/main.go
@@ -175,65 +175,6 @@ func queryBundle(bundle *semantic.GroupedBundleDataMaps, path string, line, char
 	return nil
 }
 
-func makeExistenceFunc(root string) pathexistence.GetChildrenFunc {
-	return func(ctx context.Context, dirnames []string) (map[string][]string, error) {
-		cmd := exec.Command("git", append([]string{"-C", root, "ls-tree", "--name-only", "HEAD"}, cleanDirectoriesForLsTree(dirnames)...)...)
-		out, err := cmd.CombinedOutput()
-		if err != nil {
-			fmt.Printf("git ls-tree failed with output %v\n", string(out))
-			fmt.Printf("args were %v\n", strings.Join(dirnames, ", "))
-		}
-
-		return parseDirectoryChildren(dirnames, strings.Split(string(out), "\n")), nil
-	}
-}
-
-func readBundle(dumpID int, root string) (*semantic.GroupedBundleDataMaps, error) {
-	root, err := filepath.Abs(root)
-	if err != nil {
-		fmt.Printf("Couldn't get absolute path of root %v\n", root)
-		return nil, err
-	}
-	dumpPath := path.Join(root, "dump.lsif")
-	getChildrenFunc := makeExistenceFunc(root)
-
-	gitRoot, err := gitRoot(root)
-	if err != nil {
-		return nil, err
-	}
-
-	relRoot, err := filepath.Rel(gitRoot, root)
-	// workaround: filepath.Rel returns a path starting with '../' if gitRoot and root are equal
-	if gitRoot == root {
-		relRoot = ""
-	}
-
-	file, err := os.Open(dumpPath)
-	if err != nil {
-		fmt.Printf("Couldn't open file %v\n", dumpPath)
-		return nil, err
-	}
-	defer file.Close()
-
-	bundle, err := conversion.Correlate(context.Background(), file, dumpID, relRoot, getChildrenFunc)
-	if err != nil {
-		fmt.Println("Conversion failed")
-		return nil, err
-	}
-
-	return semantic.GroupedBundleDataChansToMaps(bundle), nil
-}
-
-func gitRoot(path string) (string, error) {
-	cmd := exec.Command("git", "-C", path, "rev-parse", "--show-toplevel")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		fmt.Printf("git rev-parse --show-toplevel failed with output %v\n", string(out))
-		return "", err
-	}
-	return strings.Split(string(out), "\n")[0], nil
-}
-
 // finds all documents which have definition edges pointing into the argument list
 func documentsReferencing(bundle *semantic.GroupedBundleDataMaps, paths []string) (_ []string, err error) {
 	pathMap := map[string]struct{}{}
@@ -266,65 +207,4 @@ func documentsReferencing(bundle *semantic.GroupedBundleDataMaps, paths []string
 	}
 
 	return pathsReferencing, nil
-}
-
-// cleanDirectoriesForLsTree sanitizes the input dirnames to a git ls-tree command. There are a
-// few peculiarities handled here:
-//
-//   1. The root of the tree must be indicated with `.`, and
-//   2. In order for git ls-tree to return a directory's contents, the name must end in a slash.
-func cleanDirectoriesForLsTree(dirnames []string) []string {
-	var args []string
-	for _, dir := range dirnames {
-		if dir == "" {
-			args = append(args, ".")
-		} else {
-			if !strings.HasSuffix(dir, "/") {
-				dir += "/"
-			}
-			args = append(args, dir)
-		}
-	}
-
-	return args
-}
-
-// parseDirectoryChildren converts the flat list of files from git ls-tree into a map. The keys of the
-// resulting map are the input (unsanitized) dirnames, and the value of that key are the files nested
-// under that directory. If dirnames contains a directory that encloses another, then the paths will
-// be placed into the key sharing the longest path prefix.
-func parseDirectoryChildren(dirnames, paths []string) map[string][]string {
-	childrenMap := map[string][]string{}
-
-	// Ensure each directory has an entry, even if it has no children
-	// listed in the gitserver output.
-	for _, dirname := range dirnames {
-		childrenMap[dirname] = nil
-	}
-
-	// Order directory names by length (biggest first) so that we assign
-	// paths to the most specific enclosing directory in the following loop.
-	sort.Slice(dirnames, func(i, j int) bool {
-		return len(dirnames[i]) > len(dirnames[j])
-	})
-
-	for _, path := range paths {
-		if strings.Contains(path, "/") {
-			for _, dirname := range dirnames {
-				if strings.HasPrefix(path, dirname) {
-					childrenMap[dirname] = append(childrenMap[dirname], path)
-					break
-				}
-			}
-		} else {
-			// No need to loop here. If we have a root input directory it
-			// will necessarily be the last element due to the previous
-			// sorting step.
-			if len(dirnames) > 0 && dirnames[len(dirnames)-1] == "" {
-				childrenMap[""] = append(childrenMap[""], path)
-			}
-		}
-	}
-
-	return childrenMap
 }
